refactor #47: split tag to a single app
