+++ backend/articles/views.py
@@ -13,9 +13,10 @@ from rest_framework.permissions import (
 
 from utils.common import paginate_data, ResponseObject
 from utils.decorators import exception_handler_wrapper, construct_response
-from .models import Article, Tag, TagMap
+from .models import Article
 from .filters import TagFilterBackend, ArticleFilterBackend
 from .serializers import ArticleSerializer, TagSerializer
+from tags.models import Tag, TagMap
 
 
 class TagList(GenericAPIView):
